refactoring package for bs
