+++ src/main/java/binarysearch/Search2DMatrix.java
@@ -1,4 +1,4 @@
-package algorithms;
+package binarysearch;
 
 /**
  * https://leetcode.com/problems/search-a-2d-matrix/
